Add query needed for /api/users/:userId/decks API
Move ListDecksByUser to ListUserDecks
