+++ backend/sql/query/decks.sql
@@ -87,42 +87,3 @@ AND   NOT EXISTS (
     WHERE l2.deck_id = l.deck_id
     AND   l2.created_at > l.created_at
 );
-
--- name: ListDecksByUser :many
-SELECT
-    d.id,
-    d.name,
-    d.description,
-    d.user_id,
-    l.league_id,
-    (SELECT (
-        (
-            SELECT COUNT(*)
-            FROM duels dl
-            WHERE dl.deck_1_id = d.id
-        ) + (
-            SELECT COUNT(*)
-            FROM duels dl
-            WHERE dl.deck_2_id = d.id
-        )
-    )) AS num_duel,
-    (SELECT (
-        (
-            SELECT COUNT(*)
-            FROM duels dl
-            WHERE dl.deck_1_id = d.id AND dl.result = 1
-        ) + (
-            SELECT COUNT(*)
-            FROM duels dl
-            WHERE dl.deck_2_id = d.id AND dl.result = 2
-        )
-    )) AS num_win
-FROM decks d
-LEFT JOIN league_decks l ON d.id = l.deck_id
-WHERE d.user_id = $1 AND d.deleted_at IS NULL
-AND   NOT EXISTS (
-    SELECT 1
-    FROM league_decks l2
-    WHERE l2.deck_id = l.deck_id
-    AND   l2.created_at > l.created_at
-);
