+++ backend/sql/query/users.sql
@@ -109,3 +109,52 @@ FROM duels dl2
 WHERE dl2.user_2_id = $1
     AND dl2.confirmed_at IS NOT NULL
 ORDER BY created_at;
+
+-- name: ListUserDecks :many
+SELECT
+    d.id,
+    d.name,
+    d.description,
+    l.league_id,
+    (SELECT (
+        (
+            SELECT COUNT(*)
+            FROM duels dl
+            WHERE dl.deck_1_id = d.id
+                AND dl.confirmed_at IS NOT NULL
+                AND dl.deleted_at IS NULL
+        ) + (
+            SELECT COUNT(*)
+            FROM duels dl
+            WHERE dl.deck_2_id = d.id
+                AND dl.confirmed_at IS NOT NULL
+                AND dl.deleted_at IS NULL
+        )
+    )) AS num_duel,
+    (SELECT (
+        (
+            SELECT COUNT(*)
+            FROM duels dl
+            WHERE dl.deck_1_id = d.id
+                AND dl.result = 1
+                AND dl.confirmed_at IS NOT NULL
+                AND dl.deleted_at IS NULL
+        ) + (
+            SELECT COUNT(*)
+            FROM duels dl
+            WHERE dl.deck_2_id = d.id
+                AND dl.result = 2
+                AND dl.confirmed_at IS NOT NULL
+                AND dl.deleted_at IS NULL
+        )
+    )) AS num_win
+FROM decks d
+LEFT JOIN league_decks l ON d.id = l.deck_id
+WHERE d.user_id = $1
+    AND d.deleted_at IS NULL
+    AND NOT EXISTS (
+        SELECT 1
+        FROM league_decks l2
+        WHERE l2.deck_id = l.deck_id
+            AND l2.created_at > l.created_at
+    );
