+++ backend/sqlc/db/decks.sql.go
@@ -185,87 +185,6 @@ func (q *Queries) ListDecks(ctx context.Context) ([]ListDecksRow, error) {
 	return items, nil
 }
 
-const listDecksByUser = `-- name: ListDecksByUser :many
-SELECT
-    d.id,
-    d.name,
-    d.description,
-    d.user_id,
-    l.league_id,
-    (SELECT (
-        (
-            SELECT COUNT(*)
-            FROM duels dl
-            WHERE dl.deck_1_id = d.id
-        ) + (
-            SELECT COUNT(*)
-            FROM duels dl
-            WHERE dl.deck_2_id = d.id
-        )
-    )) AS num_duel,
-    (SELECT (
-        (
-            SELECT COUNT(*)
-            FROM duels dl
-            WHERE dl.deck_1_id = d.id AND dl.result = 1
-        ) + (
-            SELECT COUNT(*)
-            FROM duels dl
-            WHERE dl.deck_2_id = d.id AND dl.result = 2
-        )
-    )) AS num_win
-FROM decks d
-LEFT JOIN league_decks l ON d.id = l.deck_id
-WHERE d.user_id = $1 AND d.deleted_at IS NULL
-AND   NOT EXISTS (
-    SELECT 1
-    FROM league_decks l2
-    WHERE l2.deck_id = l.deck_id
-    AND   l2.created_at > l.created_at
-)
-`
-
-type ListDecksByUserRow struct {
-	ID          int32         `json:"id"`
-	Name        string        `json:"name"`
-	Description string        `json:"description"`
-	UserID      string        `json:"user_id"`
-	LeagueID    sql.NullInt32 `json:"league_id"`
-	NumDuel     int32         `json:"num_duel"`
-	NumWin      int32         `json:"num_win"`
-}
-
-func (q *Queries) ListDecksByUser(ctx context.Context, userID string) ([]ListDecksByUserRow, error) {
-	rows, err := q.db.QueryContext(ctx, listDecksByUser, userID)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	var items []ListDecksByUserRow
-	for rows.Next() {
-		var i ListDecksByUserRow
-		if err := rows.Scan(
-			&i.ID,
-			&i.Name,
-			&i.Description,
-			&i.UserID,
-			&i.LeagueID,
-			&i.NumDuel,
-			&i.NumWin,
-		); err != nil {
-			return nil, err
-		}
-		items = append(items, i)
-	}
-	if err := rows.Close(); err != nil {
-		return nil, err
-	}
-	if err := rows.Err(); err != nil {
-		return nil, err
-	}
-	return items, nil
-}
-
 const updateDeck = `-- name: UpdateDeck :exec
 UPDATE decks
 SET name = $2,
